Cleaned code and added config, Part 4 of 4
I realise we keep committing this change back and forth, but it's better/easier for the options menu to work with so that the user can say "full screen, please" and that gets saved to a file such that it's full screen every time they start it up. A significant number (not sure if most) of games do not bother with allowing the user to resize the window with dragging for the reason above and more.